requirements.txt
Flask
google-generativeai
Werkzeug
python-dotenv
